+++ Documents/Java/workspace/Projet_Tuteure_S2_bis/src/personnage/PJ.java
@@ -28,7 +28,7 @@ public class PJ extends Personnage {
 		this.PA = 6;
 	}
 	
-	public PJ(double x, double y, int direction, boolean moving, Animation[] animations,String nom,double PV, double PX, int force, int adresse, int resistance, Niveaux initiative, Niveaux attaque, Niveaux esquive, Niveaux defense, Niveaux degats, List<Objets> inventaire, Equipement equipement){
+	public PJ(float x, float y, int direction, boolean moving, Animation[] animations,String nom,double PV, double PX, int force, int adresse, int resistance, Niveaux initiative, Niveaux attaque, Niveaux esquive, Niveaux defense, Niveaux degats, List<Objets> inventaire, Equipement equipement){
 		super(x,y,direction,moving, animations, nom, PV, PX,initiative, attaque, esquive, defense, degats);
 		this.force = force;
 		this.adresse = adresse;
